Foi feito algus ajustes

+++ util/email.py
@@ -1,10 +1,9 @@
 from django.core.mail import send_mail
 from django.conf import settings
-from decimal import Decimal
 
 def send_email_for_user(
         name_product: str, 
-        new_product_price: Decimal, 
+        new_product_price: float, 
         recipient: str) -> None:
 
     subject: str = 'Preço Alterado!'
